refactor(middleware): group service middleware
