+++ coap/middleware/doc.go
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Package middleware provides tracing, logging and metrics middleware
-// for Magistrala Journal service.
+// for Magistrala COAP service.
 //
 // For more details about tracing instrumentation for Magistrala messaging refer
 // to the documentation at https://docs.magistrala.abstractmachines.fr/tracing/.
