+++ users/middleware/tracing.go
@@ -1,7 +1,7 @@
 // Copyright (c) Abstract Machines
 // SPDX-License-Identifier: Apache-2.0
 
-package tracing
+package middleware
 
 import (
 	"context"
@@ -21,8 +21,8 @@ type tracingMiddleware struct {
 	svc    users.Service
 }
 
-// New returns a new group service with tracing capabilities.
-func New(svc users.Service, tracer trace.Tracer) users.Service {
+// TracingMiddleware returns a new users service with tracing capabilities.
+func TracingMiddleware(svc users.Service, tracer trace.Tracer) users.Service {
 	return &tracingMiddleware{tracer, svc}
 }
