--- users/tracing/doc.go
@@ -1,12 +1,0 @@
-// Copyright (c) Abstract Machines
-// SPDX-License-Identifier: Apache-2.0
-
-// Package tracing provides tracing instrumentation for SuperMQ Users service.
-//
-// This package provides tracing middleware for SuperMQ Users service.
-// It can be used to trace incoming requests and add tracing capabilities to
-// SuperMQ Users service.
-//
-// For more details about tracing instrumentation for SuperMQ messaging refer
-// to the documentation at https://docs.supermq.abstractmachines.fr/tracing/.
-package tracing
