+++ ws/middleware/doc.go
@@ -1,8 +1,8 @@
 // Copyright (c) Abstract Machines
 // SPDX-License-Identifier: Apache-2.0
 
-// Package middleware provides tracing, logging and metrics middleware
-// for Magistrala Journal service.
+// Package middleware provides logging, metrics and tracing middleware
+// for Magistrala WebSockets service.
 //
 // For more details about tracing instrumentation for Magistrala messaging refer
 // to the documentation at https://docs.magistrala.abstractmachines.fr/tracing/.
